Upload images to app, replace popups with messageboxes

+++ Bierbank/ViewModel/BierDetailModel.cs
@@ -1,17 +1,25 @@
 ﻿using Bierbank.Extensions;
 using Bierbank.Model;
+using Microsoft.Win32;
 using System;
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
+using System.Diagnostics;
+using System.IO;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
+using System.Windows;
 using System.Windows.Input;
+using System.Windows.Media.Imaging;
 
 namespace Bierbank.ViewModel
 {
     public class BierDetailModel : BaseViewModel
     {
+        //variabelen
+        private string fullPath;
+
         //het biertje waarvoor we details weergeven
         private Biertjes selectedBiertje;
         public Biertjes SelectedBiertje
@@ -60,6 +68,7 @@ namespace Bierbank.ViewModel
         public ICommand WijzigenCommand { get; set; }
         public ICommand DeleteCommand { get; set; }
         public ICommand WeergevenCommand { get; set; }
+        public ICommand ImageUploadCommand { get; set; }
 
         public BierDetailModel()
         {
@@ -82,6 +91,7 @@ namespace Bierbank.ViewModel
             WijzigenCommand = new BaseCommand(UpdateBiertje);
             DeleteCommand = new BaseCommand(VerwijderBiertje);
             WeergevenCommand = new BaseCommand(BierNoteDetailWeergeven);
+            ImageUploadCommand = new BaseCommand(UploadenFoto);
         }
 
         //details over de gekozen biernote weergeven
@@ -93,11 +103,35 @@ namespace Bierbank.ViewModel
             }
         }
 
+        //foto uploaden
+        private void UploadenFoto()
+        {
+            OpenFileDialog fileDialog = new OpenFileDialog();
+            fileDialog.Title = "Selecteer een foto";
+            fileDialog.Filter = "All supported graphics|*.jpg;*.jpeg;*.png|" +
+            "JPEG (*.jpg;*.jpeg)|*.jpg;*.jpeg|" +
+            "Portable Network Graphic (*.png)|*.png";
+
+            if (fileDialog.ShowDialog() == true)
+            {
+                fullPath = fileDialog.FileName;
+                string path = System.IO.Path.GetFileName(fullPath);
+                SelectedBiertje.Image = GetDestinationPath(path, @"Images");
+            }
+        }
+
         //bier aanpassen
         private void UpdateBiertje()
         {
             BierDataService ds = new BierDataService();
             ds.UpdateBiertje(SelectedBiertje);
+
+            //image toevoegen aan de app
+            string destinationPath = SelectedBiertje.Image;
+            if (!File.Exists(destinationPath))
+            {
+                File.Copy(fullPath, destinationPath, true);
+            }
 
             //refresh
             BierenHerladen();
@@ -106,8 +140,11 @@ namespace Bierbank.ViewModel
         //bier verwijderen
         private void VerwijderBiertje()
         {
-            BierDataService ds = new BierDataService();
-            ds.DeleteBiertje(SelectedBiertje);
+            if (MessageBox.Show("Bent u hier zeker van", "verwijderen", MessageBoxButton.YesNo) == MessageBoxResult.Yes)
+            {
+                BierDataService ds = new BierDataService();
+                ds.DeleteBiertje(SelectedBiertje);
+            }
 
             //refresh
             BierenHerladen();
@@ -119,6 +156,18 @@ namespace Bierbank.ViewModel
             BierDataService ds = new BierDataService();
             ObservableCollection<Biertjes> biertjes = ds.GetBiertjes();
             Messenger.Default.Send<ObservableCollection<Biertjes>>(biertjes);
+        }
+
+        //pad om foto in op te slagen vinden
+        private static String GetDestinationPath(string file, string folder)
+        {
+            //root pad van de app vinden
+            String root = System.IO.Path.GetDirectoryName(Process.GetCurrentProcess().MainModule.FileName);
+
+            //naar gekozen folder gaan
+            root = String.Format(root + @"\{0}\" + file, folder);
+
+            return root;
         }
     }
 }
